+++ javaproperties/reading.py
@@ -1,7 +1,8 @@
-from   __future__ import unicode_literals
+from   __future__  import unicode_literals
+from   collections import namedtuple
 import re
-from   six        import binary_type, StringIO, BytesIO, unichr
-from   .util      import CONTINUED_RGX, ascii_splitlines
+from   six         import binary_type, StringIO, BytesIO, text_type, unichr
+from   .util       import CONTINUED_RGX, ascii_splitlines
 
 def load(fp, object_pairs_hook=dict):
     """
@@ -33,7 +34,9 @@ def load(fp, object_pairs_hook=dict):
     :raises InvalidUEscapeError: if an invalid ``\\uXXXX`` escape sequence
         occurs in the input
     """
-    return object_pairs_hook((k,v) for k,v,_ in parse(fp) if k is not None)
+    return object_pairs_hook(
+        (kv.key, kv.value) for kv in parse(fp) if isinstance(kv, KeyValue)
+    )
 
 def loads(s, object_pairs_hook=dict):
     """
@@ -65,50 +68,185 @@ def loads(s, object_pairs_hook=dict):
     fp = BytesIO(s) if isinstance(s, binary_type) else StringIO(s)
     return load(fp, object_pairs_hook=object_pairs_hook)
 
-COMMENT_OR_BLANK_RGX = re.compile(r'^[ \t\f]*(?:[#!]|\r?\n?$)')
+TIMESTAMP_RGX = re.compile(
+    r'\A[ \t\f]*[#!][ \t\f]*'
+    r'(?:Sun|Mon|Tue|Wed|Thu|Fri|Sat)'
+    r' (?:Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)'
+    r' (?:[012][0-9]|3[01])'
+    r' (?:[01][0-9]|2[0-3]):[0-5][0-9]:(?:[0-5][0-9]|6[01])'
+    r' (?:[A-Za-z_0-9]{3})?'
+    r' [0-9]{4,}'
+    r'[ \t\f]*\r?\n?\Z'
+)
+
+class PropertiesElement(object):
+    """
+    .. versionadded:: 0.7.0
+
+    Superclass of objects returned by `parse()`
+    """
+    __slots__ = ()
+
+
+class Comment(PropertiesElement, namedtuple('Comment', 'source')):
+    """
+    .. versionadded:: 0.7.0
+
+    Subclass of `PropertiesElement` representing a comment
+
+    .. attribute:: source
+
+        The raw, unmodified input line (including trailing newlines)
+    """
+
+    __slots__ = ()
+
+    @property
+    def value(self):
+        """
+        Returns the contents of the comment, with leading whitespace, the
+        comment marker, and trailing newline removed
+        """
+        s = self.source.lstrip(' \t\f')
+        if s.startswith(('#', '!')):
+            s = s[1:].lstrip(' \t\f')
+        return s.rstrip('\r\n')
+
+    @property
+    def source_stripped(self):
+        """
+        Like `source`, but with the final trailing newline (if any) removed
+        """
+        return self.source.rstrip('\r\n')
+
+    def is_timestamp(self):
+        """
+        Returns `True` iff the comment's value appears to be a valid timestamp
+        as produced by |date_tostring2|_
+
+        .. |date_tostring2| replace:: Java 8's ``Date.toString()``
+        .. _date_tostring2: https://docs.oracle.com/javase/8/docs/api/java/util/Date.html#toString--
+        """
+        return bool(TIMESTAMP_RGX.match(self.source))
+
+
+class Whitespace(PropertiesElement, namedtuple('Whitespace', 'source')):
+    """
+    .. versionadded:: 0.7.0
+
+    Subclass of `PropertiesElement` representing a line that is either empty or
+    contains only whitespace (and possibly some line continuations)
+
+    .. attribute:: source
+
+        The raw, unmodified input line (including trailing newlines)
+    """
+
+    __slots__ = ()
+
+    @property
+    def source_stripped(self):
+        """
+        Like `source`, but with the final trailing newline and line
+        continuation (if any) removed
+        """
+        s = self.source.rstrip('\r\n')
+        if CONTINUED_RGX.search(s):
+            s = s[:-1]
+        return s
+
+
+class KeyValue(PropertiesElement, namedtuple('KeyValue', 'key value source')):
+    """
+    .. versionadded:: 0.7.0
+
+    Subclass of `PropertiesElement` representing a key-value entry
+
+    .. attribute:: key
+
+        The entry's key, after processing escape sequences
+
+    .. attribute:: value
+
+        The entry's value, after processing escape sequences
+
+    .. attribute:: source
+
+        The concatenation of the raw, unmodified lines in the input (including
+        trailing newlines) from which the key and value were extracted
+    """
+
+    __slots__ = ()
+
+    @property
+    def source_stripped(self):
+        """
+        Like `source`, but with the final trailing newline and line
+        continuation (if any) removed
+        """
+        s = self.source.rstrip('\r\n')
+        if CONTINUED_RGX.search(s):
+            s = s[:-1]
+        return s
+
+
+COMMENT_RGX = re.compile(r'^[ \t\f]*[#!]')
+BLANK_RGX = re.compile(r'^[ \t\f]*\r?\n?\Z')
 SEPARATOR_RGX = re.compile(r'(?<!\\)(?:\\\\)*([ \t\f]*[=:]|[ \t\f])[ \t\f]*')
 
-def parse(fp):
+def parse(src):
     """
-    Parse the contents of the `~io.IOBase.readline`-supporting file-like object
-    ``fp`` as a simple line-oriented ``.properties`` file and return a
-    generator of ``(key, value, original_lines)`` triples for every entry in
-    ``fp`` (including duplicate keys) in order of occurrence.  The third
-    element of each triple is the concatenation of the unmodified lines in
-    ``fp`` (including trailing newlines) from which the key and value were
-    extracted.  The generator also includes comments and blank/all-whitespace
-    lines found in ``fp``, one triple per line, with the first two elements of
-    the triples set to `None`.  This is the only way to extract comments from a
-    ``.properties`` file with this library.
+    Parse the given data as a simple line-oriented ``.properties`` file and
+    return a generator of `PropertiesElement` objects representing the
+    key-value pairs (as `KeyValue` objects), comments (as `Comment` objects),
+    and blank lines (as `Whitespace` objects) in the input in order of
+    occurrence.
 
-    ``fp`` may be either a text or binary filehandle, with or without universal
-    newlines enabled.  If it is a binary filehandle, its contents are decoded
-    as Latin-1.
+    If the same key appears multiple times in the input, a separate `KeyValue`
+    object is emitted for each entry.
+
+    ``src`` may be a text string, a bytes string, or a text or binary
+    filehandle/file-like object supporting the `~io.IOBase.readline` method
+    (with or without universal newlines enabled).  Bytes input is decoded as
+    Latin-1.
+
 
     .. versionchanged:: 0.5.0
         Invalid ``\\uXXXX`` escape sequences will now cause an
         `InvalidUEscapeError` to be raised
 
-    :param fp: the file from which to read the ``.properties`` document
-    :type fp: file-like object
-    :rtype: generator of triples of text strings
+    .. versionchanged:: 0.7.0
+        `parse()` now accepts strings as input, and it now returns a generator
+        of custom objects instead of triples of strings
+
+    :param src: the ``.properties`` document
+    :type src: string or file-like object
+    :rtype: Iterable[PropertiesElement]
     :raises InvalidUEscapeError: if an invalid ``\\uXXXX`` escape sequence
         occurs in the input
     """
-    def lineiter():
-        while True:
-            ln = fp.readline()
-            if isinstance(ln, binary_type):
-                ln = ln.decode('iso-8859-1')
-            if ln == '':
-                return
-            for l in ascii_splitlines(ln):
-                yield l
-    liter = lineiter()
+    if isinstance(src, binary_type):
+        liter = iter(ascii_splitlines(src.decode('iso-8859-1')))
+    elif isinstance(src, text_type):
+        liter = iter(ascii_splitlines(src))
+    else:
+        def lineiter():
+            while True:
+                ln = src.readline()
+                if isinstance(ln, binary_type):
+                    ln = ln.decode('iso-8859-1')
+                if ln == '':
+                    return
+                for l in ascii_splitlines(ln):
+                    yield l
+        liter = lineiter()
     for source in liter:
         line = source
-        if COMMENT_OR_BLANK_RGX.match(line):
-            yield (None, None, source)
+        if COMMENT_RGX.match(line):
+            yield Comment(source)
+            continue
+        elif BLANK_RGX.match(line):
+            yield Whitespace(source)
             continue
         line = line.lstrip(' \t\f').rstrip('\r\n')
         while CONTINUED_RGX.search(line):
@@ -117,13 +255,17 @@ def parse(fp):
             source += nextline
             line += nextline.lstrip(' \t\f').rstrip('\r\n')
         if line == '':  # series of otherwise-blank lines with continuations
-            yield (None, None, source)
+            yield Whitespace(source)
             continue
         m = SEPARATOR_RGX.search(line)
         if m:
-            yield (unescape(line[:m.start(1)]),unescape(line[m.end():]),source)
+            yield KeyValue(
+                unescape(line[:m.start(1)]),
+                unescape(line[m.end():]),
+                source,
+            )
         else:
-            yield (unescape(line), '', source)
+            yield KeyValue(unescape(line), '', source)
 
 SURROGATE_PAIR_RGX = re.compile(r'[\uD800-\uDBFF][\uDC00-\uDFFF]')
 ESCAPE_RGX = re.compile(r'\\(u.{0,4}|.)')